Minimal scene handler should be ready, but objects aren't drawing yet

+++ 4. cube-rewrite/src/opengl/loop.hpp
@@ -7,7 +7,7 @@ namespace gl_loop {
 extern float deltaTime;
 extern float lastFrame;
 
-int init_gl(int width, int height, const char* title);
+int init_gl(float width, float height, const char* title);
 void framebuffer_size_callback(GLFWwindow *window, int width, int height);
 void mouse_callback(GLFWwindow *window, double xpos, double ypos);
 void process_input(GLFWwindow *window);
